+++ Magic/CHANGELOG.md
@@ -17,6 +17,7 @@
 
  - Fix some issues with 1.10 compatibility having to do with the Guava version
  - Add "shops" section in messages.yml, for localizing default text in shops
+ - Add "target_name" spell parameter, for only targeting an entity with a specific name.
 
 ## 6.9.9